refactor(FileFullscreenModal): update fileHeaderHeight import paths and improve error handling in PDFViewer
