+++ src/script/components/FileFullscreenModal/PdfViewer/PdfViewer.styles.ts
@@ -19,7 +19,7 @@
 
 import {CSSObject} from '@emotion/react';
 
-import {fileHeaderHeight} from '../common/fileHeaderHeight/fileHeaderHeight';
+import {fileHeaderHeight} from '../common/fileHeaderHeight';
 
 export const wrapperStyles: CSSObject = {
   height: `calc(100% - ${fileHeaderHeight}px)`,
