+++ src/script/components/FileFullscreenModal/common/fileHeaderHeight.ts
